==> CODE/0 Data Exploration and Transformation.ipynb <==
import math
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import random

d = pd.read_csv('intersection congestion.csv')

d.columns

d.iloc[random.sample(range(1,800000),5), 0:15]

d.isna().sum()

d['NameMissing'] = 0
d.loc[(pd.isna(d['EntryStreetName'])==True) & (pd.isna(d['ExitStreetName'])==False), 'NameMissing'] = 1
d.loc[(pd.isna(d['EntryStreetName'])==False) & (pd.isna(d['ExitStreetName'])==True), 'NameMissing'] = 2
d.loc[(pd.isna(d['EntryStreetName'])==True) & (pd.isna(d['ExitStreetName'])==True), 'NameMissing'] = 3
d['NameMissing'].value_counts()

# Intersection ID has duplicate in different cities. Create a unique ID.
d['InterIdNew'] = d['City']+'_'+d['IntersectionId'].astype(str)

d['TurnRoute'] = d['EntryHeading'] + '_' + d['ExitHeading']

# List of intersections which has missing value
InterMissingList = d.loc[d['NameMissing']>0, 'InterIdNew'].unique()
print(len(InterMissingList))
InterMissingList

#d.loc[d['NameMissing']>0, 'InterIdNew'].value_counts()

def fill_enter(enter, turn, store):
    if pd.isna(enter) and turn in store:
        return store[turn][0]
    else:
        return enter

def fill_exit(exit, turn, store):
    if pd.isna(exit) and turn in store:
        return store[turn][1]
    else:
        return exit

for id in InterMissingList:
    store = {}
    data = d.loc[d['InterIdNew']==id, ['EntryStreetName', 'ExitStreetName', 'TurnRoute']]
    for i in range(data.shape[0]):
        row = data.iloc[i,:]
        if row.TurnRoute not in store and pd.isna(row.EntryStreetName)==False and pd.isna(row.ExitStreetName)==False:
            store[row.TurnRoute] = [row.EntryStreetName, row.ExitStreetName]
    d.loc[(d['InterIdNew']==id) & (pd.isna(d['EntryStreetName'])), 'EntryStreetName'] = d.loc[(d['InterIdNew']==id) & (pd.isna(d['EntryStreetName'])), :].apply(lambda x: fill_enter(x['EntryStreetName'], x['TurnRoute'], store), axis=1)
    d.loc[(d['InterIdNew']==id) & (pd.isna(d['ExitStreetName'])), 'ExitStreetName'] = d.loc[(d['InterIdNew']==id) & (pd.isna(d['ExitStreetName'])), :].apply(lambda x: fill_exit(x['ExitStreetName'], x['TurnRoute'], store), axis=1)

d.isna().sum()

d['NameMissing2'] = 0
d.loc[(pd.isna(d['EntryStreetName'])==True) & (pd.isna(d['ExitStreetName'])==False), 'NameMissing2'] = 1
d.loc[(pd.isna(d['EntryStreetName'])==False) & (pd.isna(d['ExitStreetName'])==True), 'NameMissing2'] = 2
d.loc[(pd.isna(d['EntryStreetName'])==True) & (pd.isna(d['ExitStreetName'])==True), 'NameMissing2'] = 3
d['NameMissing2'].value_counts()

# List of intersections which has missing value
InterMissingList2 = d.loc[d['NameMissing2']>0, 'InterIdNew'].unique()
print(len(InterMissingList2))
print(len(d['InterIdNew'].unique()))
InterMissingList2

# driving route recognition
direction_degree = {'NW':315, 'SE':135, 'NE':45, 'SW':225, 'E':90, 'W':270, 'S':180, 'N':0}
d['EntryHeadingDegree'] = d['EntryHeading'].apply(lambda x: direction_degree[x])
d['ExitHeadingDegree'] = d['ExitHeading'].apply(lambda x: direction_degree[x])
d['TurnDegree'] = d['ExitHeadingDegree'] - d['EntryHeadingDegree']
d['TurnDegree'].unique()

d_agg = d.groupby(['InterIdNew','ExitHeading'])['TurnDegree'].count()
InterStruc = d_agg.unstack()
InterStruc.fillna(0, inplace=True)
InterStruc[InterStruc>0] = 1
InterStruc['DirectionNum'] = InterStruc.apply(lambda x: x.sum(), axis=1)
InterStruc.reset_index(level=0, inplace=True)
print(InterStruc.sample(10))
print(InterStruc['DirectionNum'].value_counts())

d_agg = d.groupby(['InterIdNew','TurnRoute'])['TurnDegree'].count()
RountStruc = d_agg.unstack()
RountStruc.fillna(0, inplace=True)
RountStruc[RountStruc>0] = 1
RountStruc['RountNum'] = RountStruc.apply(lambda x: x.sum(), axis=1) 
RountStruc.reset_index(level=0, inplace=True)
print(RountStruc.sample(10))
print(RountStruc['RountNum'].value_counts())

inter = pd.merge(InterStruc[['InterIdNew', 'DirectionNum']], RountStruc[['InterIdNew', 'RountNum']], how='left', left_on='InterIdNew', right_on='InterIdNew')
print(inter.sample(10))
print(inter.groupby(['DirectionNum', 'RountNum'])['InterIdNew'].count())

d['TurnCategory'] = 'Straight'
d.loc[((d['TurnDegree']<=135) & (d['TurnDegree']>45)) | ((d['TurnDegree']<=135-360) & (d['TurnDegree']>45-360)), 'TurnCategory'] = 'Right'
d.loc[(d['TurnDegree']==180) | (d['TurnDegree']==180-360), 'TurnCategory'] = 'Uturn'
d.loc[((d['TurnDegree']<315) & (d['TurnDegree']>=225)) | ((d['TurnDegree']<315-360) & (d['TurnDegree']>=225-360)), 'TurnCategory'] = 'Left'

# time window aggregation
d['TimeSlot'] = 'Midnight(22:00-07:00)'
d.loc[d['Hour'].isin([7,8,9]), 'TimeSlot'] = 'MorningBusy(07:00-10:00)'
d.loc[d['Hour'].isin([16,17,18,19]), 'TimeSlot'] = 'EveningBusy(16:00-20:00)'
d.loc[d['Hour'].isin([20,21]), 'TimeSlot'] = 'NormalNight(20:00-22:00)'
d.loc[d['Hour'].isin([10,11,12,13,14,15]), 'TimeSlot'] = 'NormalDay(10:00-16:00)'

d1 = d.groupby(['InterIdNew', 'Weekend', 'TimeSlot', 'TurnCategory'])['TotalTimeStopped_p80'].median().unstack().reset_index()
d2 = d.groupby(['InterIdNew'])['Latitude', 'Longitude'].median()
data = pd.merge(left=d1, right=d2, how='inner', left_on='InterIdNew', right_on='InterIdNew')
data.sample(10)

# ignore Uturn
del data['Uturn']
data.sample(10)

data.to_csv('ProgressDataOutputArchive/data_agg.csv')
