==> CODE/6 Spatial Clustering Visualization.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
from scipy.spatial import Voronoi, voronoi_plot_2d
import pandas as pd
import geopandas
import pysal as ps
import json
import geojson
import folium

def AMOEBA_map(address):
    path = 'ProgressDataOutputArchive/AMOEBA/DataPrepare/'+address
    jsonfile = path+'/'+address+'_json.json'
    with open(jsonfile) as f:
        geo_data = json.load(f)
    
    datafile = 'ProgressDataOutputArchive/AMOEBA/Results/'+address+'_AMOEBA.csv'
    data = pd.read_csv(datafile)
    data['id'] = data.index+1
    data['id'] = data['id'].apply(str)
    data = data[['id', 'cluster']]
    
    m = folium.Map(location=[33.757586, -84.386598],zoom_start=14)

    folium.Choropleth(
        geo_data=geo_data,
        name='choropleth',
        data=data,
        columns=['id', 'cluster'],
        key_on='feature.id',
        fill_color='YlGn',
        fill_opacity=0.4,
        line_opacity=0.2,
    ).add_to(m)

    folium.LayerControl().add_to(m)
    
    m.save('ProgressDataOutputArchive/AMOEBA/Results/'+address+'_AMOEBA.html')

AMOEBA_map('Weekday_EveningBusy')
AMOEBA_map('Weekday_Midnight')
AMOEBA_map('Weekday_Morningbusy')
AMOEBA_map('Weekday_Normalday')
AMOEBA_map('Weekday_Normalnight')
AMOEBA_map('Weekend_Eveningbusy')
AMOEBA_map('Weekend_Midnight')
AMOEBA_map('Weekend_Morningbusy')
AMOEBA_map('Weekend_Normalday')
AMOEBA_map('Weekend_Normalnight')
